支持新疆工作 查验判定 申请表  等等优化

+++ app/src/main/java/com/seatrend/xj/electricbicyclesalesystem/createview/ClipperView.java
@@ -372,7 +372,12 @@ public class ClipperView extends View {
                 }
             }
         } else {
-            if (bottom != mViewH - defaultDis - mHWDis) {
+//            showLog("bottom = "+bottom);
+//            showLog("mViewH = "+mViewH);
+//            showLog("defaultDis = "+defaultDis);
+//            showLog("mHWDis = "+mHWDis);
+//            showLog("top = "+top);
+            if ((int)bottom != (int)(mViewH - defaultDis - mHWDis)) {
                 if (top > defaultDis + mHWDis) {
                     top = (int) (y - dt);
                     if (top < defaultDis + mHWDis) {
